+++ demo-projects/todo/CHANGELOG.md
@@ -1,3 +1,5 @@
+# @keystone/demo-project-todo
+
 # @keystone-alpha/demo-project-todo
 
 ## 1.1.18