+++ demo-projects/todo/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @keystonejs/demo-project-todo
 
+## 5.0.0
+
+### Major Changes
+
+- [`7b4ed362`](https://github.com/keystonejs/keystone-5/commit/7b4ed3623f5774d7783c39962bfa1ce97938e310) [#1821](https://github.com/keystonejs/keystone-5/pull/1821) Thanks [@jesstelford](https://github.com/jesstelford)! - Release @keystonejs/\* packages (つ＾ ◡ ＾)つ
+
+  - This is the first release of `@keystonejs/*` packages (previously `@keystone-alpha/*`).
+  - All packages in the `@keystone-alpha` namespace are now available in the `@keystonejs` namespace, starting at version `5.0.0`.
+  - To upgrade your project you must update any `@keystone-alpha/*` dependencies in `package.json` to point to `"@keystonejs/*": "^5.0.0"` and update any `require`/`import` statements in your code.
+
+### Patch Changes
+
+- Updated dependencies [[`7b4ed362`](https://github.com/keystonejs/keystone-5/commit/7b4ed3623f5774d7783c39962bfa1ce97938e310)]:
+  - @keystonejs/adapter-mongoose@5.0.0
+  - @keystonejs/app-admin-ui@5.0.0
+  - @keystonejs/app-graphql@5.0.0
+  - @keystonejs/app-static@5.0.0
+  - @keystonejs/fields@5.0.0
+  - @keystonejs/keystone@5.0.0
+
 # @keystone-alpha/demo-project-todo
 
 ## 1.1.18